+++ client/pages/[...not-found].tsx
@@ -1,15 +1,18 @@
 import React from 'react'
 
-import { GetServerSideProps } from 'next'
+import type { GetServerSidePropsResult } from 'next'
 import { useTranslation } from 'next-i18next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import { NextSeo } from 'next-seo'
 
+import { setLocale } from '@/api'
+import { wrapper } from '@/api/store'
 import AppLayout from '@/components/app-layout'
 import Menu from '@/components/app-layout/Menu'
+import { LocaleType } from '@/tools/types'
 
 export default function CatchAllPage() {
-    const { i18n, t } = useTranslation('common')
+    const { i18n, t } = useTranslation()
 
     return (
         <AppLayout>
@@ -35,13 +38,14 @@ export default function CatchAllPage() {
     )
 }
 
-export const getServerSideProps: GetServerSideProps = async (context) => {
-    const locale = context.locale || 'en'
+export const getServerSideProps = wrapper.getServerSideProps(
+    (store) =>
+        async (context): Promise<GetServerSidePropsResult<object>> => {
+            const locale: LocaleType = (context.locale as LocaleType) ?? 'ru'
+            const translations = await serverSideTranslations(locale, ['common'])
 
-    return {
-        props: {
-            ...(await serverSideTranslations(locale, ['common'])),
-            notFound: true
+            store.dispatch(setLocale(locale))
+
+            return { props: { ...translations } }
         }
-    }
-}
+)