+++ server/app/Libraries/VisualCrossingAPILibrary.php
@@ -97,7 +97,10 @@ class VisualCrossingAPILibrary
         ];
 
         try {
-            $response = $this->httpClient->request('GET', self::API_URL . getenv('app.lat') . ',' . getenv('app.lon') . '/' . $endpoint, ['query' => $params]);
+            $response = $this->httpClient->request('GET', self::API_URL . getenv('app.lat') . ',' . getenv('app.lon') . '/' . $endpoint, [
+                'query'   => $params,
+                'timeout' => 30,
+            ]);
             return json_decode($response->getBody(), true);
         } catch (Exception $e) {
             log_message('error', 'VisualCrossing API request error: {e}', ['e' => $e->getMessage()]);